egg/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking setup.py
#
certifi==2022.12.7
    # via minio
click==8.1.3
    # via edx-backup (setup.py)
minio==7.1.13
    # via edx-backup (setup.py)
pendulum==2.1.2
    # via edx-backup (setup.py)
pydantic==1.10.4
    # via edx-backup (setup.py)
python-dateutil==2.8.2
    # via pendulum
pytzdata==2020.1
    # via pendulum
six==1.16.0
    # via python-dateutil
terminaltables==3.1.10
    # via edx-backup (setup.py)
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.14
    # via minio
